plugins/dashboard/requirements.txt
flask>=2.0.0
pyyaml>=5.0.0
